Rename DateInputDateEntry to DateInputEntry
