+++ apps/docs/src/routes/base/date-input/examples/disabled.tsx
@@ -5,13 +5,13 @@ export default component$(() => {
   return (
     <DateInput.Root>
       <DateInput.Label>Permanently Disabled</DateInput.Label>
-      <DateInput.DateEntry date="2000-12-25" disabled={true}>
+      <DateInput.Entry date="2000-12-25" disabled={true}>
         <DateInput.Month />
         <DateInput.Separator separator="/" />
         <DateInput.Day />
         <DateInput.Separator separator="/" />
         <DateInput.Year />
-      </DateInput.DateEntry>
+      </DateInput.Entry>
     </DateInput.Root>
   );
 });
