+++ apps/docs/src/routes/base/date-input/examples/placeholder.tsx
@@ -6,13 +6,13 @@ export default component$(() => {
   return (
     <DateInput.Root>
       <DateInput.Label>Enter your date of birth:</DateInput.Label>
-      <DateInput.DateEntry>
+      <DateInput.Entry>
         <DateInput.Month placeholder="Month" />
         <DateInput.Separator separator="/" />
         <DateInput.Day placeholder="Day" />
         <DateInput.Separator separator="/" />
         <DateInput.Year placeholder="Year" />
-      </DateInput.DateEntry>
+      </DateInput.Entry>
     </DateInput.Root>
   );
 });
