+++ apps/docs/src/routes/base/date-input/examples/reactive.tsx
@@ -10,13 +10,13 @@ export default component$(() => {
     <div class="date-input-container">
       <DateInput.Root class="date-input-root-col">
         <DateInput.Label>Party like it's:</DateInput.Label>
-        <DateInput.DateEntry bind:date={selectedDate}>
+        <DateInput.Entry bind:date={selectedDate}>
           <DateInput.Month />
           <DateInput.Separator separator="/" />
           <DateInput.Day />
           <DateInput.Separator separator="/" />
           <DateInput.Year />
-        </DateInput.DateEntry>
+        </DateInput.Entry>
       </DateInput.Root>
       <p>
         Bound date: <span class="external-value">{selectedDate.value}</span>
