+++ apps/docs/src/routes/base/date-input/examples/toggle-disabled.tsx
@@ -10,13 +10,13 @@ export default component$(() => {
     <div class="date-input-container">
       <DateInput.Root class="date-input-root-col">
         <DateInput.Label>My date input</DateInput.Label>
-        <DateInput.DateEntry disabled={isDisabled.value} date="1998-11-01">
+        <DateInput.Entry disabled={isDisabled.value} date="1998-11-01">
           <DateInput.Month />
           <DateInput.Separator separator="/" />
           <DateInput.Day />
           <DateInput.Separator separator="/" />
           <DateInput.Year />
-        </DateInput.DateEntry>
+        </DateInput.Entry>
       </DateInput.Root>
 
       <div class="date-input-button-container">
