+++ apps/docs/src/routes/base/date-input/examples/value-based.tsx
@@ -13,13 +13,13 @@ export default component$(() => {
     <div class="date-input-container">
       <DateInput.Root class="date-input-root-col">
         <DateInput.Label>Choose your date:</DateInput.Label>
-        <DateInput.DateEntry date={selectedDate.value} onChange$={handleChange$}>
+        <DateInput.Entry date={selectedDate.value} onChange$={handleChange$}>
           <DateInput.Month showLeadingZero />
           <DateInput.Separator separator="/" />
           <DateInput.Day showLeadingZero />
           <DateInput.Separator separator="/" />
           <DateInput.Year />
-        </DateInput.DateEntry>
+        </DateInput.Entry>
       </DateInput.Root>
 
       <p>
