+++ libs/components/src/date-input/date-input-entry-context.tsx
@@ -2,11 +2,11 @@ import { type Signal, createContextId } from "@builder.io/qwik";
 import type { ISODate } from "../calendar/types";
 import type { DateSegment } from "./types";
 
-export const dateInputDateEntryContextId = createContextId<DateInputDateEntryContext>(
-  "qds-date-input-date-entry-context"
+export const dateInputEntryContextId = createContextId<DateInputEntryContext>(
+  "qds-date-input-entry-context"
 );
 
-export type DateInputDateEntryContext = {
+export type DateInputEntryContext = {
   dateSig: Signal<ISODate | null>;
   dayOfMonthSegmentSig: Signal<DateSegment>;
   monthSegmentSig: Signal<DateSegment>;
