+++ libs/components/src/date-input/date-input-year.tsx
@@ -1,6 +1,6 @@
 import { component$, useContext } from "@builder.io/qwik";
 import { getNextIndex } from "@kunai-consulting/qwik-utils";
-import { dateInputDateEntryContextId } from "./date-input-date-entry-context";
+import { dateInputEntryContextId } from "./date-input-entry-context";
 import { DateInputSegment } from "./date-input-segment";
 import type { PublicDateInputSegmentProps } from "./types";
 
@@ -14,7 +14,7 @@ export const DateInputYearBase = component$(
     showLeadingZero = false,
     ...otherProps
   }: PublicDateInputSegmentProps) => {
-    const context = useContext(dateInputDateEntryContextId);
+    const context = useContext(dateInputEntryContextId);
     const segmentSig = context.yearSegmentSig;
 
     return (
