+++ libs/components/src/date-input/index.ts
@@ -1,5 +1,5 @@
 export { DateInputRoot as Root } from "./date-input-root";
-export { DateInputDateEntry as DateEntry } from "./date-input-date-entry";
+export { DateInputEntry as Entry } from "./date-input-entry";
 export { DateInputHiddenInput as HiddenInput } from "./date-input-hidden-input";
 export { DateInputLabel as Label } from "./date-input-label";
 export { DateInputSegment as Segment } from "./date-input-segment";
